Initial Bayesian neural network code.
This includes functions for easily
* generating normal priors + variational distributions for all weights of a `torch.nn.Module`,
* making a BNN using a model + priors,
* making a guide with the variational distributions,
* training a BNN using SVI,
* getting an array of predictions of samples from a BNN,
* selecting a batch of new data as a subset of given data using PDTS or a greedy approach (for a BNN or just an NN),
* evaluating Bayesian optimization with a given model/acquisition function by determining how long it takes to find the top 1% of inputs.

Note that all of the acquisition functions expect to be passed a set of potential inputs and then acquire their batch as a subset of that; this only works for finite (and not too enormous) input spaces.
Also note that in the PDTS paper the model is trained using probabilistic backpropagation; using SVI, so far at least, has yielded inferior results.