Added HSIC acqusition function + faster implementation.
Added a dummy model + random acqusition function for the uniform sampling baseline to `bayesian_opt` so that it conforms to the same method of running the other acquisition functions. The new HSIC acquisition function is there and uses the same BNN model + training as PDTS and greedy BNN.

The script to run Bayesian optimization is now in its own file `run_bo`. It, and somewhat inside of the functions in `bayesian_opt` as well, now logs to Comet ML (there will probably be errors in some places if this can't be done).

Training the BNN now supports early stopping (on the train loss, not a validation set).

Simplified the RQ kernel code and made it use the same device for new tensors as its input; similarly for `total_hsic` (at least the log version). Also added functions which enable efficient computation of HSIC for several points in parallel which are each under consideration to add to an existing batch.

Added a loading function for the saved fraction_best dictionaries to utils and a different plotting method.