Calibration: better Gaussian support, initialize w/ regressor.
Now you can directly pass in the means and standard deviations of the Guassian at each point instead of only samples from these (useful in the case when the CDF truly is Gaussian, not just approximated so).
A calibration regressor can be given when initializing a CDF; this lets you train a regressor on a CDF with train labels and then use that regressor on a CDF with test data.
`get_cdf_probs` now supports multiple query points per CDF even in the Gaussian case.